# docs/specs/openapi.yaml
openapi: 3.1.0
info:
  title: KK Studio Unified API
  version: 1.0.0
  summary: REST + OpenAPI 3.1 contract for the server-backed API and current billing routes.
  description: |
    This contract freezes the first wave of external and internal-facing APIs used by KK Studio.
    All routes return a unified JSON envelope and use typed DTOs through the current services/api/ backend.
servers:
  - url: https://api.kkai.plus
    description: Main API / BFF
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/billing/credits/balance:
    get:
      tags: [Billing/Credits API]
      summary: Get current credit balance
      operationId: getCreditBalance
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XClientVersion'
      responses:
        '200':
          description: Credit balance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditBalanceResponseEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/billing/credits/transactions:
    get:
      tags: [Billing/Credits API]
      summary: List credit transactions for the current user
      operationId: listCreditTransactions
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XClientVersion'
        - in: query
          name: transactionType
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 100
      responses:
        '200':
          description: Credit transaction list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditTransactionListResponseEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/billing/credits/debit:
    post:
      tags: [Billing/Credits API]
      summary: Debit user credits for a business action
      operationId: debitCredits
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XClientVersion'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DebitCreditsRequest'
      responses:
        '200':
          description: Credit debit result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebitCreditsResponseEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/billing/credits/refunds:
    post:
      tags: [Billing/Credits API]
      summary: Refund a completed debit transaction
      operationId: refundCredits
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XClientVersion'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefundCreditsRequest'
      responses:
        '200':
          description: Credit refund result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefundCreditsResponseEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/billing/recharge-submissions:
    post:
      tags: [Billing/Credits API]
      summary: Create a new manual recharge submission
      operationId: createRechargeSubmission
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XClientVersion'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount]
              properties:
                amount:
                  type: number
                currencyCode:
                  type: string
                paymentChannel:
                  type: string
                manualProvider:
                  type: string
      responses:
        '201':
          description: Recharge submission created
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    required: [submission]
                    properties:
                      submission:
                        $ref: '#/components/schemas/RechargeSubmission'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/billing/recharge-submissions/{submissionId}/proof:
    post:
      tags: [Billing/Credits API]
      summary: Upload payment proof for a recharge submission
      operationId: uploadRechargeProof
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XClientVersion'
        - name: submissionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                transferReferenceLast4:
                  type: string
                note:
                  type: string
      responses:
        '200':
          description: Recharge proof uploaded
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    required: [submission]
                    properties:
                      submission:
                        $ref: '#/components/schemas/RechargeSubmission'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/billing/recharge-submissions/{submissionId}/mark-paid:
    post:
      tags: [Billing/Credits API]
      summary: Mark a recharge submission as paying
      operationId: markRechargePaid
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XClientVersion'
        - name: submissionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Recharge submission marked as paying
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    required: [submission]
                    properties:
                      submission:
                        $ref: '#/components/schemas/RechargeSubmission'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    XRequestId:
      in: header
      name: X-Request-Id
      required: true
      schema:
        type: string
        format: uuid
    XClientVersion:
      in: header
      name: X-Client-Version
      required: false
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponseEnvelope'
    Unauthorized:
      description: Authentication is required or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponseEnvelope'
...
    NotFound:
      description: Resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponseEnvelope'
    Conflict:
      description: Business conflict detected
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponseEnvelope'
...
  schemas:
    Meta:
      type: object
      required: [requestId, timestamp]
      properties:
        requestId:
          type: string
          format: uuid
        clientVersion:
          type: string
        timestamp:
          type: string
          format: date-time
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
...
    ErrorResponseEnvelope:
      type: object
      required: [success, error, meta]
      properties:
        success:
          type: boolean
          const: false
        error:
          $ref: '#/components/schemas/ErrorObject'
        meta:
          $ref: '#/components/schemas/Meta'
...
    CreditBalance:
      type: object
      required: [accountId, userId, balance, frozenBalance, updatedAt]
      properties:
        accountId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        balance:
          type: integer
        frozenBalance:
          type: integer
        updatedAt:
          type: string
          format: date-time
    CreditBalanceResponseEnvelope:
      type: object
      required: [success, data, meta]
      properties:
        success:
          type: boolean
          const: true
        data:
          $ref: '#/components/schemas/CreditBalance'
        meta:
          $ref: '#/components/schemas/Meta'
...
    CreditTransactionListResponseEnvelope:
      type: object
      required: [success, data, meta]
      properties:
        success:
          type: boolean
          const: true
        data:
          $ref: '#/components/schemas/CreditTransactionList'
        meta:
          $ref: '#/components/schemas/Meta'
    DebitCreditsRequest:
      type: object
      required: [businessRefType, businessRefId, creditAmount, idempotencyKey]
      properties:
        businessRefType:
          type: string
        businessRefId:
          type: string
        creditAmount:
          type: integer
          minimum: 1
        modelCode:
          type: string
        idempotencyKey:
          type: string
...
    DebitCreditsResponseEnvelope:
      type: object
      required: [success, data, meta]
      properties:
        success:
          type: boolean
          const: true
        data:
          $ref: '#/components/schemas/DebitCreditsResponse'
        meta:
          $ref: '#/components/schemas/Meta'
    RefundCreditsRequest:
      type: object
      required: [transactionId]
      properties:
        transactionId:
          type: string
          format: uuid
        reason:
          type: string
...
    RefundCreditsResponseEnvelope:
      type: object
      required: [success, data, meta]
      properties:
        success:
          type: boolean
          const: true
        data:
          $ref: '#/components/schemas/RefundCreditsResponse'
        meta:
          $ref: '#/components/schemas/Meta'
...
    RechargeSubmission:
      type: object
      required: [submissionId, userId, amount, currencyCode, status, createdAt]
      properties:
        submissionId:
          type: string
        userId:
          type: string
        amount:
          type: number
        baseAmount:
          type: number
        serviceFee:
          type: number
        payableAmount:
          type: number
        baseCredits:
          type: integer
        bonusCredits:
          type: integer
        creditAmount:
          type: integer
        creditsPerUnit:
          type: integer
        currencyCode:
          type: string
        paymentChannel:
          type: string
        manualProvider:
          type: string
          nullable: true
        transferReferenceLast4:
          type: string
          nullable: true
        note:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        paymentMarkedAt:
          type: string
          format: date-time
          nullable: true
        submittedAt:
          type: string
          format: date-time
          nullable: true
        reviewedAt:
          type: string
          format: date-time
          nullable: true
        reviewActorUserId:
          type: string
          nullable: true
...